move shopping cart model to order-flow directory
